Extract selfie spinner into component
